Support windows drive letters
Drive letters "C:\..." mess with the current approach of using colon
as separator for various inline probe configs.

This is an attempt to make this a bit more robust. We do not special
case this behavior when running on windows, since it's too easy to
break this on other platforms.

Change-Id: I6279ffa36f5a0b38ef1c5cd8871b57b793f7f5c8
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/4774458
Commit-Queue: Camillo Bruni <cbruni@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>