Fix downloading webdriver when /tmp is a different file system
Some linux distributions like to have /tmp mounted as a different
filesystem, via tmpfs. However, this means we cannot use Path.rename()
to move a file from/to the /tmp directory, as it uses os.rename() which
does not support moving files across different disks.

Instead, use shutil.move(). Note that passing a Path to it is only
supported in python 3.9, so we have to use os.fspath() to convert it.

Change-Id: I1e9f306cd76b4a4a84ab2ab69e50ad6686b14032
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/4774459
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Camillo Bruni <cbruni@chromium.org>