Change Flags.__iter__ to return strings instead of items
Flags is a dual-purpose object where it implements most of the
dictionary interface, but when using it we want it to behave like
a list.

This CL changes the __iter__ implementation to no longer return
dict items but strings instead. This is more convenient for most
usages.

Flags.to_dict() now covers the missing dict conversion which is used
for testing.

Change-Id: I2c21f97cdd0794b79e5b27fde3e4a4e250f4d4fd
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5562783
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Camillo Bruni <cbruni@chromium.org>