Improve Android tap action
When an app is docked (as in with a foldable) there are potentially
adjustments that must be made to the element position as reported by
chrome.

Additionally, handle the case where clank does not report the correct
devicePixelRatio for zoomed pages.

Tests: modified speedometer test that uses a 'tap' to click the start
button tested on: emulated pixel fold, emulated generic phone, emulated
tablet.

Change-Id: Ia96b86bca406d81552d4de2a0dac277c34f3c10c
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5732816
Commit-Queue: Kameron Lutes <kalutes@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>