crossbench: support opening multiple tabs in loading test
This change adds an option to open the given urls in multiple tabs.

It will be used for testing how many tabs can be kept alive at the same time.

Eventually, we target to create a stress benchmark test that opens
multiple windows and tabs.

It can be run via:
        ./cb.py loading --browser='adb:chrome-app' \
          --splash-screen=none \
          --urls="cnn,facebook,amazon" \
          --tabs='multiple'

Bug=b:358122440
Change-Id: I7c43801df9e5d132dfd897c32e1e3d3e1840e7a3
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5758235
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Suhua Lei <lsuhua@google.com>