Add a new result type for TraceProcessor compatible files
Longer term we want Perfetto traces to be the source of truth for a run.
For now lets simplify how traces compatible with TraceProcessor are
collected and aggregated.

The TraceProcessor probe is not really a probe, as it just deals with
result processing. Add some special handling to enfore that the probe is
used correctly (has to be the first probe specified)

Change-Id: Ic352303bfc54cfe518ed79ff1c90b3d6479b066d
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5791300
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Carlos Caballero Grolimund <carlscab@google.com>