Fix double-initialization of logging
The default description of the loading benchmark requires a list of
story names. PresetPageLoadBenchmark has to parse its config to get
story names, and the parsing code in its turn emits log lines. This
causes a premature initialization of logging, and as a result
everything is logged twice later on.

This CL replaces the default description to avoid initializing logging
too early.

Change-Id: I3e04e4c289c6ebca609bcd4454f88a94eff07b32
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5845488
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>
Reviewed-by: Carlos Caballero Grolimund <carlscab@google.com>