Support running WPR on device
Web loading timings on an Android device depend a lot on the connection
speed between the device and the workstation. This makes loading
benchmark scores inconstistent between different setups.

This CL makes it possible to run WPR on the device thus eliminating
the variance in loading speeds.

Change-Id: I4c49f983ddddd33d2484d1bc3b664e9823c718fd
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5876290
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>