loading android click fail if element has no area
An element with no with or height means we either didn't find an element
with that selector, or the one we did find doesn't actually take up any
pixels on the screen.

Either way, we shouldn't be injecting a click since it will not be on
the intended target.

Bug: 366102372
Test: poetry run cb loading --browser adb:chrome --action-runner android
--page-config http://gpaste/6506536664039424 -- -- --request-desktop-sites

Change-Id: Iec12a18ae2a1712700d96ed121220590084b8b00
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5855002
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Charles Dick <cwd@google.com>