loading switch_tab action
Switches the tab that actions are executed on. WebDriver doesn't provide
a list of window titles or URLs, so all switching is done by cycling one
tab/window at a time. If title or URL filters are provided, keep cycling
until a matching window is found or the timeout is reached.

Example, switch to a window that is created by a click action:
  {action: "click", selector: "..."},
  {action: "switch_tab", title: "^new window title"},
If the window isn't immediately available, all windows are cycled until
one with a matching title is found.

Example, switch to the next window:
  {action: "switch_tab"},
If only one tab exists, it will check for new windows until timeout is
reached.

Example, switch to the 5th tab created, but only succeed if the title
starts with 'A':
  {action: "switch_tab", tab_index: 5, title: "^A"},

Parameters
  tab_index - Optional int index of the tab to switch to
  title - Optional regexp to match tab/window titles.
  url - Optional regexp to match tab/window URLs.
  timeout - How long to retry if no matching tab is found.

Bug: b:365886816
Test: pytest, loading page_config http://gpaste/6506536664039424
Change-Id: Id61a1caf4c75b435c67875cc2f24ea0c8990c034
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5850058
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Charles Dick <cwd@google.com>