Use live URL to replay WPR in the press benchmarks.
WPR archive files are recorded using a live URL, and the same URL must be used for playback. However,  path and url cannot both be set in a network_config. Currently, press_benchmark.py assumes a local file server is being used if a custom URL isn't provided. This is problematic because it overwrites the original WPR URL. To address this, we need to explicitly check if the WPR network is using a local file server. If it is, we need to generate a matching URL for it.

Bug: 369705307
Change-Id: I76a3ce633b5bc04e4db80e785822292c63e58abc
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5905925
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Farid (Mojtaba) Faridzad <faridzad@google.com>