Basic support for clearing WPR after running the setup code
We don't always want to record the login/setup code in a WPR as it might
contain login information. We use the not-yet implemented wpr
`command-clear` command. This behavior will be configured with the newly
added `record_setup` option for the WPR-probe.

Add more flags to alter which blocks we run:
- --skip-login skips the login block
- --skip-setup skips the setup block

Change-Id: I682e297bbae2fb6f1469d575de02ec0289fe849f
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5899858
Reviewed-by: Farid (Mojtaba) Faridzad <faridzad@google.com>
Commit-Queue: Camillo Bruni <cbruni@chromium.org>