dump_html probe and loading action
Introduce DumpHtmlProbe, which can dump the HTML of a page upon request
but by default does nothing.

HTML is dumped on loading test failure if the probe is running.

Introduce DumpHtmlAction, which dumps the HTML of a page. Intended for
debugging during development.

Bug: b:353623546
Test: poetry run cb loading
--probe-config <http://gpaste/5705768895774720>
--page-config <http://gpaste/5503431677837312>
Check that HTML dumps are made for the action and the failure of the
wait_for_element action.

Change-Id: I32d1255c5d8d567d3c75f588de24db507e1d1641
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5912574
Commit-Queue: Charles Dick <cwd@google.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>