chromeos-ssh auto setup chromedriver
This change allows a missing or 0 port in a chromeos-ssh driver config.
If so, chromedriver is automatically started and stopped on the chromeos
device, and its port is forwarded with SSH. The RemoteWebDriver is then
created with the forwarded port.

With this change, there is not need to do any remote configuration or
setup of a ChromeOS dut before running a test

Bug: b:372797374
Test: pytest, poetry run cb loading --browser-config <<see below>>
{
  "browsers": {
    "chromeos-chrome": {
      "path": "/opt/google/chrome/chrome",
      "driver": {
        "type": "chromeos-ssh",
        "settings": {
          "host": "dut.hostname",
          "ssh_port": 22,
          "ssh_user": "root"
        }
      }
    }
  }
}

Change-Id: Ie296c2b882182c7aceb2f96e370d7dda802f9eee
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/5823454
Reviewed-by: Saswat Padhi <spadhi@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Charles Dick <cwd@google.com>