Write custom error printer
Python's default exception hook prints the outermost exception last.

  Frame 3: outer.py parent()
  Frame 4: outer.py child()
  InnerException: message

  Frame 1: inner.py parent()
  Frame 2: inner.py child()
  OuterException: message

This is practical if you only care about the outermost exception.
But reading the stack traces for the inner exceptions becomes rather
confusing since it's not in the full order you would expect when
only looking at the inner exception.

The custom error printer chnages this to the following:

 Frame 1: inner.py parent()
 Frame 2: inner.py child()
 OuterException: message

 Frame 3: outer.py parent()
 Frame 4: outer.py child()
 InnerException: message

This means that reading the full stack from the a single exception
is the same as one which was rethrown with a "from" exception.

Change-Id: I5247641eebed2437e27b8a9586cf2a9a4bfb46d3
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6021672
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Camillo Bruni <cbruni@chromium.org>