More robust ConfigObject path handling
Ignore all OSErrors when deciding whether a parsed input string
is a separate config file or not.

Depending on the value we get, we might trigger an OSError
on is_file / is_dir checks due to the paths being long or
accidentally being inaccessible.

Given that the fallback is to pass along the values to
parse_str, we can safely ignore those errors.

Change-Id: Icc4f4da0872cfa9bcfb7deb654a9fdb586497dca
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6070026
Commit-Queue: Camillo Bruni <cbruni@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>