loading remove tab switch before get target=_self
We want to be able to load 20 tabs, and then go back to each tab and
navigate to a blank page so that the LCP metrics get logged.

Remving the tab switch on a get action allows us to do this, and it is
not needed by memory benchmarks any more (which was the reason it was
added).

See conversation:
https://chat.google.com/room/AAAAtLxUatw/tN1ZCQDSZ04

Bug: b:354089900
Change-Id: I4e84007e96938be1cd22d739f77b4e0f29f244e7
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6151375
Commit-Queue: Charles Dick <cwd@google.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>