loading click action verify
The verify argument on loading test click actions is a selector that is
run after the click. If the selector doesn't match anything within the
action's timeout then the action fails.

This means that misclicks due to incorrect timing or location of
injected input events will be caught on the action that failed, and not
at some later point, if ever, in the test.

Bug: b:352451846

Change-Id: I21777d3b954e2152523e41f646de864ab09cb2b1
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6088664
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Charles Dick <cwd@google.com>