MacOS profiling: Support profiling rendering process only
So far MacOS used system wide profiling unconditionally. This change
adds support for the 'target' parameter of the profiling probe for
MacOS. Currently supported are 'SYSTEM_WIDE' and
'RENDERING_PROCESS_ONLY'.

Note that 'RENDERING_PROCESS_ONLY' is only supported on Chromium-based
browsers as it requires an extension for retrieving the renderer pid.

Note that for now the default value of `target` is unsupported on Mac,
so in order to use this we need to specify a supported target:
```
--probe='profiling:{target:"RENDERER_PROCESS_ONLY"}'
```

Fixed: 390993105
Change-Id: I6b5a8f3473cfe0367e3222ea462ffd5fb403c058
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6180625
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>