Use Platform-specific signals independently from the host platform
Currently using posix-specific signals for an attached android
device on a windows host is broken. The posix signals are not
accessible since we relied on python's native signal library which
only exposes windows signals in this case.

crossbench now maintains it's own copy of various platform-specific
signals independent from the host-platform we're running on.

As part of this change we now rely on platform methods directly
for process handling.

- proc_helper methods now use an explicit platform argument
- Move _IGNORED_PROCESS_EXCEPTIONS to proc_helper
- Expose Platform.wait_and_kill and Platform.wait_and_terminate
- Make sure RemotePopen has the correct returncode
- Collect output for remote PosixPlatform.send_signal
- TsProxy classes now take an explicit platform

Fixed: 391570177
Change-Id: I581c5d70b64c96fd238652b45083f1e2e5337c24
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6190461
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Camillo Bruni <cbruni@chromium.org>