Prepare FlagsGroupConfig for more complex cli flags
Add FlagsGroupConfig.parse_args which will handle extra browser
args and in the future will split it into separate flag groups.

The final goal is to be able to quickly specify inline groups
for testing chrome features without the need of a separate browser
config file

./cb.py sp3 -- \
   LABEL1: --enable-features=FOO \
   LABEL2: --disable-features=FOO

This CL is the first step to add the basic APIs and tests.
In a follow-up CL we will add the separator / label parsing and
the actual group splitting logic.

Change-Id: I12cdd3df2bf7efd4f3aa4877a371af2b4c5f3d9d
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6261887
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Camillo Bruni <cbruni@chromium.org>