Allow comparison for BrowserVersions in the same class hierarchy
Previously we only allow comparing BrowserVersion object if the right
object is a subclass of the left object.

With this CL comparisons are allowed if either is a subclass of the
other.

This fixes a min-version check for the profiling probe on linux
where a chrome version check is used for feature detection.

Drive-by-fix:
- Improve version comparison error message

Bug: 397252110
Change-Id: Ifb5f4792e9290d233ec8932baa7c25f7a257b053
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6276578
Commit-Queue: Camillo Bruni <cbruni@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>