loading add retry for click actions
web-tests docs is flaky on ChromeOS because the file menu occasionally
disappears after opening if the click happens too early in page load,
but after the complete event.

Add an attempts property to click actions that retrys the whole click
action if there is a failure.

Bug: b:407899220
Test: web-tests docs
Change-Id: If5b2eec85b551d603fb9239b8a63e56bc5ab8cc1
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6461304
Reviewed-by: Kameron Lutes <kalutes@google.com>
Commit-Queue: Charles Dick <cwd@google.com>