loading fix setup and login failure artefacts
We didn't set the info_stack on the action runner when running setup or
login. The dump_html and screenshot probes use the info_stack to name
their outputs, so this caused us to skip artefacts during login and
setup.

Set the info_stack so that we actually get results on failures during
setup and login.

Bug: b:407899220
Test: web-tests meet, but with Google login broken by changing a
selector

Change-Id: I0e3fdc174725103c452e50b30f44965a28a1c02d
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6472239
Reviewed-by: Kameron Lutes <kalutes@google.com>
Commit-Queue: Charles Dick <cwd@google.com>