Enable "d8" browser
For cli-based benchmarks we can reuse most of the code with d8.
The D8 browser is a minimal wrapper around the d8-shell
which is used to load and execute JS code.

In the future, we can support more JS-only benchmarks by extending
the URLMapper, currently we only support JetStream.

Change-Id: I00f7693b621bb5cc95e68c48ed00ff73fab611fe
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6557542
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Camillo Bruni <cbruni@chromium.org>