Swallow a broader category of exceptions in WprBase._shut_down()
There seems to be a race condition where if the connection pool is
closed before the request enters it, ConnectionError will be raised -
but the opposite order can also happen in which case ReadTimeout is
raised instead. Both are subclasses of RequestException, so with the
proposed change we cover all bases.

Bug: 393058910
Change-Id: Ib00337531ac42a0874a249b7f39289da5c896d91
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6578258
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Auto-Submit: Mateusz Zbikowski <zbikowski@google.com>
Commit-Queue: Camillo Bruni <cbruni@chromium.org>