Support fuzzy matching
Enable fuzzy matching for more config params. The following items
will auto-resolve to the closest match if difflib returns only one
item. Otherwise subset ot choices is printed:
- Benchmark names
- Probe names
- Story names

This also cleans up the cli by not directly exposing the benchmark
aliases through the argparser interface. Aliases are not handled
manually in cli.js and support fuzzy matching.

The aliases are now only printed as part of the benchmark help
or the cli help, but not explicitly listed as choices for the
subcommand. This makes error messages a bit more readable.


Change-Id: Ifbc3393f67578e006ea0b47db2df60920708c88b
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6578565
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Camillo Bruni <cbruni@chromium.org>