[LoadLine] Switch to archives with pre-injected deterministic.js
Injecting a script on-the-go can be slow and this can become a blocker
in the page loading process. To make sure LoadLine correctly measures
the browser page loading performance, we disable runtime script
injection and use an archive with pre-injected script instead.

Bump the LoadLine version since on some devices this can affect scores.

Change-Id: I9a1079f741ad32351aab9f7f5d98904aa346e835
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6634600
Reviewed-by: Victor Vianna <victorvianna@google.com>
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>