Add support for cycled accounts to secrets
When running crossbench as part of tradefed (or other lab setting) it's
necessary to have a list of accounts that all are used for tests. Add
support for random account selection from a range when specifying
accounts as part of secrets.

The specified format (%d and an int range) is consistent with what is
currently used in Android tradefed.

Change-Id: Ib8357beea07ae96ea4548ba973c953b8da8bf890
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6636623
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Kameron Lutes <kalutes@google.com>