Drop loadline2 dep on PageLoadMetrics.NavigationToLargestContentfulPaint
This trace event is Chrome-specific. Removing the dependency on it is a
first step towards making loadline2 cross-browser.

Instead of relying on the event, now the google search and docs stories
emit performance.mark() when the LCP element is added to the DOM, then
feed that timestamp to get_presentation_time() - just like other
stories. The LCP elements were found manually with DevTools.

Bug: 372457479
Change-Id: I420ea8f7b4f74653fbaac09a067c717b0ecba09e
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6646514
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Commit-Queue: Victor Vianna <victorvianna@google.com>