Make sure WPR process is killed before exiting
Port forwarding stops before atexit hooks are called, so at the moment
when we try to shutdown WPR, it's no longer possible to send commands
to it via network requests. So to avoid stale processes accumulating
on the device, we have to kill the process before exit.

Change-Id: I572e73834926131753a02e892cc78dcf7753aedd
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6650795
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>