Add 'title' option for meminfo action
Add an optional 'title' field for a meminfo action that is included in
the meminfo performance mark. This is useful for determining at what
point in a loading test a meminfo dump is from. If no title is supplied,
the current info stack is used.

Change-Id: I1dbf37a6d9e733b64c2e308e50b94333254daf4c
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6666731
Commit-Queue: Kameron Lutes <kalutes@google.com>
Reviewed-by: Charles Dick <cwd@google.com>