Fix BrowserProfilingProbe
More details in the linked bug. The probe was trying to call
Tracing.tracingComplete, which is an event, not a ChromeDevtoolsProtocol
method. This CL fixes the issue by calling Tracing.end() to end the
trace collection and then watching for the Tracing.tracingComplete
event, which contains the collected trace.

Fixed: 420743776
Change-Id: If173635cf76cac02d0af65dfe1fd6cf1e49680fa
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6592915
Commit-Queue: Victor Vianna <victorvianna@google.com>
Auto-Submit: Victor Vianna <victorvianna@google.com>
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>