browser config extensions parameter
Specify a list of extensions in the BrowserConfig. Extensions can be an
extension ID, a path to a crx file, or a path to an unpacked extension.

Extensions specified by ID are downloaded from the Chrome Web Store for
the current browser version.

Packed extensions are unpacked to a temporary folder.

If the browser platform is remote, the now unpacked extensions are
pushed to a temp folder.

Finally, extensions are provided to Chrome with --load-extension, and
feature DisableLoadExtensionCommandLineSwitch is disabled to allow
extensions to load.

Why not use Selenium options.add_extension?

Selenium pushes extensions to the browser by base-64 encoding them in
the setup request. Request bodies have a limit of 100MB, which limits
what extensions can be used. This could be fixed, but not for older
versions of Chrome.

Also, Android chromedriver does not support extensions at all. I tried
implementing support, but ran into issues pushing files to the Android
device. Chromedriver only has a socket to the adb server, so adb push is
not available and had to be re-implemented. Pushing certain files
failed. At this point I gave up and implemented everything in
Crossbench instead.

Test: web-test tab-stress.blank extensions variant on cros, android,
local

Change-Id: I760655b34ac05db6e756c6463aa88f292b91a96d
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6567345
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Charles Dick <cwd@google.com>