refactor Platform.meminfo
We should be passing lists of ProcessMeminfo because cmdline is not
guaranteed to be unique so we shouldn't key off it.

Update LinuxPlatform to use a script to collect all the meminfos, the
latency of per-process ssh commands to cat cmdline and smaps_rollup
causes meminfo actions to take minutes when lots of tabs are open.

Why not use a shell one-liner? I could not get shell escaping to work
properly on remote Linux.

Change-Id: Iab1385cfefabe0f9bf930747f0c25a44273ea371
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6702890
Commit-Queue: Charles Dick <cwd@google.com>
Reviewed-by: Kameron Lutes <kalutes@google.com>