Handle empty repetition results for ResultsSummaryProbe
Skip empty results when merging repetitions. An EmptyProbeResult is
expected when session setup fails. In this case we should fall back
to the default ProbeResult.merge behaviour of ignoring empty results.
This is the equivalent of crrev.com/c/6714025 which handled the same
for JsonResultProbe.

Bug: 393059233
Change-Id: I7f25bfe30e69e0f89b96980a5c956cc606b57264
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6771849
Commit-Queue: Mateusz Zbikowski <zbikowski@google.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>