Manage the chromedriver service log file in ChromiumBasedWebDriver
Right now _stdout_log_file is opened in one class and closed in
another which creates problems. E.g. when a webdriver session fails
to start the log file remains open.

Since this file is only used in ChromiumBasedWebDriver, it makes
sense to manage it completely within that class, and also rename
the variable to _service_log_file (since this is how it's used).

Change-Id: I81b1587026ffd1188b52e748f3fae01edffbad13
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6763617
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>