Add probe action
Add a new action type that can be used for general purpose invocation of
a specific probe.

Going forward, all actions that interact directly with probes should not
be implemented as new action classes, but as an implementation of the
invoke() method on the probe context class.

Change-Id: Ib85ebb66d431fe420630e110a2c4c201078ac2e7
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6818109
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Kameron Lutes <kalutes@google.com>