[LoadLine 1] Fix breakdown query
The breakdown query for LoadLine v1 assumed that the timestamp
used in the metric calculation is the end timestamp of a
PipelineReporter slice. In fact, it's the timestamp of a
Display::FrameDisplayed slice (see first_presentation_time.sql),
which is not guaranteed to be the same (although it almost always is).

This CL fixes the query in the case where they are not the same.

Bug: 445356773
Change-Id: Iddf7058abed04dc9f2dc80c84f9ee69e87abd966
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6951340
Reviewed-by: Victor Vianna <victorvianna@google.com>
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>