[LoadLine 2] Use existing renderer for some navigations
In real use cases, not all navigations require a new renderer, some
use an existing one. To make LoadLine 2 more representative of real-
world page load performance, we create a spare renderer to use in a
subset of pages. We do this by opening a new tab with a blank page
before the navigation.

The subset of pages is chosen to contain both fast-loading and slow-
loading pages.

Change-Id: I56931c94048587508f1aea2141043d921bd54aad
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6938876
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>
Reviewed-by: Victor Vianna <victorvianna@google.com>