Specify action when starting embedder
In latest versions of the embedder app, the action name needs to be
specified to land in the "Suggest" view and not the Home screen. This
change is also backwards compatible and works with older versions of the
embedder.

Change-Id: Ief9f65e013d21b6425f3c1459b0a997c2fb4dc47
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6964629
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Mateusz Zbikowski <zbikowski@google.com>
Auto-Submit: Mateusz Zbikowski <zbikowski@google.com>