Add `--ignore-partial-failures` option to the runner.
This new option prevents the runner from failing the entire execution
when only some runs or sessions encounter errors. When enabled, the
runner will still log the errors but will not raise a `RunnerException`
at the end, allowing the process to complete and potentially report on
the successful parts.

Bug: 432061439
Change-Id: Ib70921f8807e5a17ec462e55660b40dcf20d5f5d
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6973705
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Mateusz Zbikowski <zbikowski@google.com>