Add configuration for running setup commands in embedder benchmarks.
This change introduces `SetupCommandsConfig` to parse a dictionary of
named commands, where each command is a list of strings. The
`embedder_benchmark.py` is updated to include an
`--embedder-setup-command-config` argument, allowing users to specify
setup commands to be run before the benchmark. Tests are added to verify
the parsing of valid and invalid configurations.

Change-Id: Iaafd05597c0e8fc19ebf1e5b413869f08c2411ae
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6973460
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Mateusz Zbikowski <zbikowski@google.com>