Try all open tabs when collecting embedder metrics.
The embedder browser now iterates through all open window handles,
attempting to check the responsiveness of each. If an error occurs on a
tab, the driver is restarted, and the next tab is tried. This handles
scenarios where the initial active tab is not the one containing the
necessary JavaScript context.

Change-Id: Id7aaf8d0705ce99777e04d23121b81f5f52a8952
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/6985161
Commit-Queue: Mateusz Zbikowski <zbikowski@google.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>