Add the 'pinpoint config' command
This CL adds a new subcommand to the `pinpoint` subcommand. The new
subcommand is `config` displays a job config for an existing job by its
ID. The job config is displayed in a simplified and more structured way
than the original config. The displayed config is supposed be used for
creating new jobs.

Bug: b/455510346
Change-Id: I7c1ccd0e7462411dc63e4e5fdfaabd9bb7c46f11
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/7098921
Reviewed-by: Camillo Bruni <cbruni@chromium.org>