If the experiment git hash is not provided, set to the base git hash
The pinpoint backend can create a job if the request contains experiment
git hash equals None but the job will not be executed properly. As a
result, the job UI will be broken.

This CL uses the base git hash value for the experiment git hash if the
latter is not provided.

Bug: b/455510346
Change-Id: I95705c1ae3adf414646195820e8617f970d94078
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/7122844
Commit-Queue: Maxim Sheshukov <maximsheshukov@google.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>