Add `--filter` flag for pinpoint `bots` and `benchmarks` commands
This CL introduces a new flag `--filter` for the pinpoint subcommand.
If the flag is set, the displayed list of bots or benchmarks is filtered by case insensitive matching.

In order to reduce code duplication, this CL creates a base class for filtering the results for pinpoint subcommands which display a string list e.g. `pinpoint bots`.

In order to reduce code duplication, this CL creates a base test case class MockAuthSessionMixin which mocks the `get_auth_session` function, which is commonly used in the pinpoint command.

Bug: b/457678610
Change-Id: Ib2bc8e071a151e961ea500a70a694d1c9a6e66c2
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/7132039
Commit-Queue: Maxim Sheshukov <maximsheshukov@google.com>
Reviewed-by: Patrick Thier <pthier@chromium.org>