Support URLs for the pinpoint `--job` flag
Several pinpoint subcommands require a job ID to be provided. Sometimes
it is convenient to provide a job URL or even a part of a job URL which includes the job ID.

This CL makes the usage of the `--job` flag consistent across all
pinpoint subcommands by introducing a dedicated parser and a base class
for subcommands which need the `--job` flag.

Bug: b/457674342
Change-Id: I0c8a4cc21cbb83ffeeb1870585a84ae4ee88785c
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/7198044
Commit-Queue: Maxim Sheshukov <maximsheshukov@google.com>
Reviewed-by: Patrick Thier <pthier@chromium.org>