Add a new 'pinpoint results' command for downloading job results
This CL implements a command for downloading pinpoint job results. The
command accepts a job and a directory folder. Then, it creates a
subdirectory named <date>_pinpoint_<benchmark>_<bot> and downloads an
HTML results file.

Separate subdirectories are created for each
experiment arm. If it is a try job, "base" and "exp" directories are
created. If it is a bisect job, "chromium_<commit_number>_..."
directories are created.

Separate directories are created for each attempt. This directories
follow the same order as in the Web UI and named with the same index.

Bug: b/457675058
Change-Id: I42bd58624ed749cdd59f3610ffac2d3c6d5caa98
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/7198182
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Maxim Sheshukov <maximsheshukov@google.com>