Collect Pinpoint CLI user metrics
This CL sends a random anonymous user ID and the executed command to a
BigQuery table.

The user is notified about metrics collection on the first executed
command. Then, the response is stored in a configuration file. If the
response is negative, no metrics will be collected.

Bug: b/463582182
Change-Id: I309c00470e0d6b2a8806fc72e161db750acbd065
Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/7221971
Commit-Queue: Maxim Sheshukov <maximsheshukov@google.com>
Reviewed-by: Patrick Thier <pthier@chromium.org>